add query
