+++ sql/parse/parse.go
@@ -530,7 +530,7 @@ func convertShow(ctx *sql.Context, s *sqlparser.Show, query string) (sql.Node, e
 		return node, nil
 	case "function status":
 		var filter sql.Expression
-
+		var node sql.Node
 		if s.Filter != nil {
 			if s.Filter.Filter != nil {
 				var err error
@@ -547,7 +547,14 @@ func convertShow(ctx *sql.Context, s *sqlparser.Show, query string) (sql.Node, e
 			}
 		}
 
-		var node sql.Node = plan.NewShowFunctionStatus()
+		node, err := Parse(ctx, "select routine_schema as `Db`, routine_name as `Name`, routine_type as `Type`,"+
+			"definer as `Definer`, last_altered as `Modified`, created as `Created`, security_type as `Security_type`,"+
+			"routine_comment as `Comment`, character_set_client, collation_connection,"+
+			"database_collation as `Database Collation` from information_schema.routines where routine_type = 'FUNCTION'")
+		if err != nil {
+			return nil, err
+		}
+
 		if filter != nil {
 			node = plan.NewFilter(filter, node)
 		}
